Move streaming regressions to OpenApiTransformer tests
